Move trend line computation to chart model
